Issue #47 - anchor invoice
* start implementing invoice controller for anchor only

* 1st running impl of anchor only for invoice

* testing - disable glide cache to pull protobufs from pstehlik repo

* testing - disable glide cache to pull protobufs from pstehlik repo

* revert travis.yml

* allow running tests but have consistent GETH node running locally

* allow running tests but have consistent GETH node running locally

* fixing quotation for test-runner

* cleanup of run scripts

* introducing logging to clean error handling up a bit

* pulling in new protobufs version instead of pstehlik's fork

+++ scripts/test_wrapper.sh
@@ -4,6 +4,9 @@ set -a
 # Setup
 local_dir="$(dirname "$0")"
 PARENT_DIR=`pwd`
+GETH_DOCKER_CONTAINER_NAME="geth-node"
+GETH_DOCKER_CONTAINER_WAS_RUNNING=`docker ps -a --filter "name=${DOCKER_CONTAINER_NAME}" --filter "status=running" --quiet`
+echo "Running: [${GETH_DOCKER_CONTAINER_WAS_RUNNING}]"
 
 ################# Run Dependencies #########################
 for path in ${local_dir}/test-dependencies/*; do
@@ -12,7 +15,7 @@ for path in ${local_dir}/test-dependencies/*; do
     echo "Executing [${path}/run.sh]"
     ${path}/run.sh
     if [ $? -ne 0 ]; then
-      exit 1
+        exit 1
     fi
 done
 ############################################################
@@ -59,14 +62,18 @@ fi
 ############################################################
 
 ################# CleanUp ##################################
-echo "Bringing GETH Daemon Down"
-docker rm -f geth-node
+if [ -n "${GETH_DOCKER_CONTAINER_WAS_RUNNING}" ]; then
+    echo "Container ${GETH_DOCKER_CONTAINER_NAME} was already running before the test setup. Not tearing it down as the assumption is that the container was started outside this context."
+else
+    echo "Bringing GETH Daemon Down"
+    docker rm -f geth-node
 
-# Cleaning extra DAG file, so we do not cache it - travis
-if [[ "X${RUN_CONTEXT}" == "Xtravis" ]];
-then
-  new_dag=`ls -ltr $DATA_DIR/.ethash/* | tail -1 | awk '{print $9}' | tr -d '\n'`
-  rm -Rf $new_dag
+    # Cleaning extra DAG file, so we do not cache it - travis
+    if [[ "X${RUN_CONTEXT}" == "Xtravis" ]];
+    then
+      new_dag=`ls -ltr $DATA_DIR/.ethash/* | tail -1 | awk '{print $9}' | tr -d '\n'`
+      rm -Rf $new_dag
+    fi
 fi
 ############################################################
